cambios

+++ src/page/Title/ListTitles.js
@@ -4,6 +4,9 @@ import { FormatterDate } from '../../scripts/FormatterDate';
 import { FaList, FaRegEdit, FaTrashAlt } from 'react-icons/fa';
 import { OverlayTrigger, Tooltip } from 'react-bootstrap';
 import { encryptAES } from '../../scripts/AES-256';
+import Swal from "sweetalert2";
+import { Toaster, toast } from 'sonner';
+
 
 function ListTitles() {
     const bodyRef = useRef();
@@ -18,36 +21,67 @@ function ListTitles() {
             setShow(false);
         } else if (user === "Admin_Library") {
             setShow(true);
+        }else{
+            window.location.href = `/`;
         }
 
         getAll()
             .then((result) => {
-                setTitles(result); // Actualiza el estado de titles
+                if (result.length === 0) {
+                    toast.warning('Ooops,No hay libros registrados');
+                    setTimeout(() => {
+                        window.location.href = `/addTitle`;
+                    }, 1000);
+                } else {
+                    setTitles(result); // Actualiza el estado de titles
+                }
             })
             .catch(() => {
-                console.log("Error al obtener los libros");
+                toast.error('Ooops,Algo salió mal');
             });
     }, []); // Pasa un arreglo vacío como segundo argumento para que useEffect se ejecute solo una vez
 
     function handleDeleteTitle(id) {
-        deleteTitle(id)
-            .then((result) => {
-                console.log(result + "Titulo eliminada exitosamente");
-                window.location.reload();
-            })
-            .catch(() => {
-                console.log("Error al eliminar la titulo");
-            });
+        Swal.fire({
+            title: '¿Estás seguro de Eliminar el Libro?',
+            text: "No podrás revertir esto.",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Sí, eliminarlo!'
+        }).then(async (result) => {
+            if (result.isConfirmed) {
+                Swal.fire(
+                    '¡Eliminado!',
+                    'El Libro ha sido eliminado.',
+                    'success'
+                )
+                deleteTitle(id)
+                    .then((result) => {
+                        window.location.reload();
+                    })
+                    .catch(() => {
+                        toast.error('Ooops,Algo salió mal');
+                    });
+            } else {
+
+                Swal.fire
+                    (
+                        'Error',
+                        'No se pudo eliminar el Libro.',
+                        'error'
+                    );
+            }
+        })
     }
 
     const handleSubmit = (e) => {
         e.preventDefault();
         const value = e.target[0].value;
-        console.log(value);
         const filteredTitles = titles.filter((element) => {
             return element.title1.toLowerCase().includes(value.toLowerCase());
         });
-        console.log(filteredTitles);
         setTitles(filteredTitles);
     }
 
@@ -59,7 +93,7 @@ function ListTitles() {
                 setTitles(result); // Actualiza el estado de titles
             })
             .catch(() => {
-                console.log("Error al obtener los libros");
+                toast.error('Ooops,Algo salió mal');
             });
     }
 
@@ -67,17 +101,20 @@ function ListTitles() {
         <div className="container">
 
             <h1>Listado de libros</h1>
-            {show ? <a href="/addTitle" className="btn btn-primary">Agregar Libro</a> : null}
             <nav className="navbar ">
                 <div className="container-fluid">
                     <form className="d-flex" role="search" onSubmit={handleSubmit}>
-                        <input className="form-control me-2 col-6" type="search" ref={seach} placeholder="Search" aria-label="Search" />
+                        <input className="form-control me-2 col-6" type="search" ref={seach} placeholder="Título" aria-label="Search" />
                         <button className="btn btn-outline-success me-2" type="submit">Buscar</button>
                         <button className="btn btn-outline-warning" onClick={handleReset}>Limpiar</button>
                     </form>
                 </div>
             </nav>
-            <div className="py-4">
+            <div className="row mb-2">
+                <div className="col-4 mb-3">
+                    <a href="/" className="btn btn-primary float-left">Regresar</a>
+                    {show ? <a href="/addTitle" className="btn btn-primary ms-2 float-rigth">Agregar Libro</a> : null}
+                </div>
                 <table className="table border shadow py-4 mb-5">
                     <thead>
                         <tr>
@@ -144,6 +181,9 @@ function ListTitles() {
                     </tbody>
                 </table>
             </div>
+            <Toaster
+                richColors
+                position='bottom-center' />
         </div>
     );
 }
